requirements.txt
asgiref==3.7.1
cfgv==3.3.1
distlib==0.3.6
Django==4.2.1
django-cors-headers==4.0.0
djangorestframework==3.14.0
filelock==3.12.0
gunicorn==20.1.0
identify==2.5.24
joblib==1.2.0
nodeenv==1.8.0
numpy==1.24.3
pandas==2.0.1
platformdirs==3.5.1
pre-commit==3.3.2
psycopg2==2.9.6
pycryptodome==3.18.0
pyflakes==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
scikit-learn==1.2.1
scipy==1.10.1
six==1.16.0
sqlparse==0.4.4
threadpoolctl==3.1.0
tzdata==2023.3
virtualenv==20.23.0
whitenoise==6.4.0
